resolved issues causing warnings

+++ src/FIMS.cpp
@@ -28,7 +28,7 @@ Type objective_function<Type>::operator()() {
       fims::Information<Type>::GetInstance();
 
     //update the parameter values
-    for(int i =0; i < information->fixed_effects_parameters.size(); i++){
+    for(size_t i =0; i < information->fixed_effects_parameters.size(); i++){
         *information->fixed_effects_parameters[i] = p[i];
     }
     model -> of = this;
